Proform Bike PFEVEX71316.0 #3157

+++ src/devices/proformbike/proformbike.cpp
@@ -1084,7 +1084,7 @@ void proformbike::characteristicChanged(const QLowEnergyCharacteristic &characte
             (newValue.at(0) != 0x00 && newValue.at(0) != 0x01) || newValue.at(1) != 0x12 ||
             (newValue.at(0) == 0x00 &&
              (newValue.at(2) != 0x01 || newValue.at(3) != 0x04 || newValue.at(4) != 0x02 || newValue.at(5) != 0x2c)) ||
-             (proform_bike_PFEVEX71316_0 && newValue.at(2) == 0xFF || newValue.at(3) == 0xFF)) {
+            (proform_bike_PFEVEX71316_0 && ((uint8_t)newValue.at(2) == 0xFF || (uint8_t)newValue.at(3) == 0xFF))) {
             return;
         }
